Move CaseNodeData back to coordination service package
